pyproject.toml
[tool.poetry]
name = "microblog API"
version = "0.1.0"
description = ""
authors = ["Aleksey Kuznetcov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
SQLAlchemy = "^2.0.30"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
async-factory-boy = "^1.0.1"
python-multipart = "^0.0.9"
aiofiles = "^23.2.1"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
black = "^24.4.2"
mypy = "^1.10.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
flake8-bugbear = "^24.4.26"
flake8-pie = "^0.16.0"
pytest-cov = "^5.0.0"
pytest-sugar = "^1.0.0"
coverage = "^7.5.3"
wemake-python-styleguide = "^0.19.2"
flake8-quotes = "^3.4.0"

[tool.coverage.report]
exclude_also = [
    "create_fake_data_bd",
    "ovr_get_current_user_by_api_key"
]

[tool.pytest.ini_options]
asyncio_mode="auto"
filterwarnings = "ignore"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"